Refactor code per PR #1232 comments
